refactor: rename history_new and header_with_proof_new by removing _new
