+++ crates/ethportal-api/src/types/content_value/history.rs
@@ -2,7 +2,7 @@ use ssz::{Decode, Encode};
 
 use crate::{
     types::{
-        content_value::ContentValue, execution::header_with_proof_new::HeaderWithProof,
+        content_value::ContentValue, execution::header_with_proof::HeaderWithProof,
         network::Subnetwork,
     },
     utils::bytes::hex_encode,
