+++ crates/ethportal-api/src/types/content_value/mod.rs
@@ -6,7 +6,7 @@ use crate::{
 pub mod beacon;
 pub mod constants;
 pub mod error;
-pub mod history_new;
+pub mod history;
 pub mod state;
 
 /// An encodable portal network content value.
