+++ crates/ethportal-api/src/types/execution/header_with_proof.rs
