+++ crates/ethportal-api/src/types/execution/mod.rs
@@ -1,7 +1,7 @@
 pub mod accumulator;
 pub mod block_body;
 pub mod header;
-pub mod header_with_proof_new;
+pub mod header_with_proof;
 pub mod receipts;
 pub mod transaction;
 pub mod withdrawal;
